+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.1.4] - 2024-03-15
+### Updated
+- Backend
+  - Added selected answer to results for display
+
+- Frontend
+  - Updated view of single test results
+
 ## [0.1.3] - 2024-03-15
 ### Updated
 - Frontend